kaihuang:I add partition feature for large train and test set in Q1 and Q2
